+++ src/app/app.routes.ts
@@ -4,21 +4,22 @@ import { LoggedInGuard } from './shared/login/logged-in.guard';
 
 import { ProfileComponent } from '././shared/user/profile.component';
 
-import { AboutComponent } from './about/index';
-import { LandingComponent } from './landing/index';
+import { AboutComponent } from './about';
+import { LandingComponent } from './landing';
+import { HomeComponent } from './home';
 
-import { LeaderEditComponent } from './leader/edit/index';
-import { LeaderListComponent } from './leader/list/index';
-import { LeaderViewComponent } from './leader/view/index';
+import { LeaderEditComponent } from './leader/edit';
+import { LeaderListComponent } from './leader/list';
+import { LeaderViewComponent } from './leader/view';
 
-import { ProjectsComponent } from './project/landing/index';
-import { ProjectEditComponent } from './project/edit/index';
-import { ProjectListComponent } from './project/list/index';
-import { ProjectViewComponent } from './project/view/index';
+import { ProjectsComponent } from './project/landing';
+import { ProjectEditComponent } from './project/edit';
+import { ProjectListComponent } from './project/list';
+import { ProjectViewComponent } from './project/view';
 
-import { TaskEditComponent } from './task/edit/index';
-import { TaskListComponent } from './task/list/index';
-import { TaskViewComponent } from './task/view/index';
+import { TaskEditComponent } from './task/edit';
+import { TaskListComponent } from './task/list';
+import { TaskViewComponent } from './task/view';
 
 //
 // The order of routes is IMPORTANT.
@@ -44,7 +45,8 @@ export const routes: Routes = [
   { path: 'profile', component: ProfileComponent, canActivate: [LoggedInGuard] },
 
   { path: 'about', component: AboutComponent },
-  { path: '', component: LandingComponent }
+  { path: '', component: HomeComponent }
+  // { path: '', component: LandingComponent }
 ];
 
 export const routing = RouterModule.forRoot(routes);